backend/app/tasks/sms.py
from celery import current_task
from app.core.celery import celery_app
import requests
from app.core.config import settings

@celery_app.task
def send_sms_task(phone_number: str, message: str):
    """Send SMS asynchronously"""
    try:
        if not settings.SMS_API_KEY:
            print(f"SMS not configured. Would send to {phone_number}: {message}")
            return {"status": "skipped", "reason": "SMS not configured"}
        
        # This is a placeholder for actual SMS service integration
        # You would integrate with services like Twilio, AWS SNS, etc.
        payload = {
            "to": phone_number,
            "message": message,
            "api_key": settings.SMS_API_KEY
        }
        
        # Simulate API call
        # response = requests.post("https://api.sms-service.com/send", json=payload)
        
        return {"status": "sent", "recipient": phone_number}
    
    except Exception as e:
        return {"status": "failed", "error": str(e)}

@celery_app.task
def send_bulk_sms_task(recipients: list, message: str):
    """Send bulk SMS asynchronously"""
    results = []
    for recipient in recipients:
        result = send_sms_task.delay(recipient, message)
        results.append(result)
    
    return {"status": "queued", "count": len(recipients)}